+++ src/lib/forms/Checkbox.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { getContext } from 'svelte';
   import type { FormColorType } from '../types';
-  import Radio, { labelClass, inputClass } from './Radio.svelte';
+  import { labelClass, inputClass } from './Radio.svelte';
   import Label from './Label.svelte';
 
   // properties forwarding
@@ -16,7 +16,17 @@
   // tinted if put in component having its own background
   let background: boolean = getContext('background');
 
-  $: {
+  // react on external group changes
+  function init(_: HTMLElement, _group: string[]) {
+    checked = checked || _group.includes(value);
+    return {
+      update(_group: string[]) {
+        checked = _group.includes(value);
+      }
+    };
+  }
+
+  function onChange() {
     // There's a bug in Svelte and bind:group is not working with wrapped checkbox
     // This workaround is taken from:
     // https://svelte.dev/repl/de117399559f4e7e9e14e2fc9ab243cc?version=3.12.1
@@ -40,6 +50,7 @@
 
 <Label class={labelClass(inline, $$props.class)} show={!!$$slots.default}>
   <input
+    use:init={group}
     type="checkbox"
     bind:checked
     on:keyup
@@ -52,6 +63,8 @@
     on:mouseenter
     on:mouseleave
     on:paste
+    on:change={onChange}
+    on:change
     {value}
     {...$$restProps}
     class={inputClass(custom, color, true, background, $$slots.default || $$props.class)} /><slot />